# lib/api-zod/src/generated/types/watchlistItem.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * TradeVision AI Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { WatchlistItemCategory } from './watchlistItemCategory';

export interface WatchlistItem {
  id: number;
  symbol: string;
  /** @nullable */
  name?: string | null;
  category: WatchlistItemCategory;
  /** @nullable */
  notes?: string | null;
  addedAt: Date;
}