# lib/api-zod/src/generated/types/watchlistItemInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * TradeVision AI Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { WatchlistItemInputCategory } from './watchlistItemInputCategory';

export interface WatchlistItemInput {
  symbol: string;
  name?: string;
  category: WatchlistItemInputCategory;
  notes?: string;
}